// src/converter/types.ts
export const FONT_FAMILY = {
	1: 'Virgil',
	2: 'Helvetica',
	3: 'Cascadia',
} as const

export declare type RoughPoint = [number, number]
export type Point = Readonly<RoughPoint>

export declare type Line = [Point, Point]
export interface Rectangle {
	x: number
	y: number
	width: number
	height: number
}

export type ChartType = 'bar' | 'line'
export type FillStyle = 'hachure' | 'cross-hatch' | 'solid'
export type FontFamily = keyof typeof FONT_FAMILY
export type FontString = string & { _brand: 'fontString' }
export type GroupId = string
export type PointerType = 'mouse' | 'pen' | 'touch'
export type StrokeSharpness = 'round' | 'sharp'
export type StrokeStyle = 'solid' | 'dashed' | 'dotted'
export type TextAlign = 'left' | 'center' | 'right'